requirements.txt
pyaml == 24.4.0
PyYAML == 6.0.1
requests == 2.31.0
pyvis == 0.3.2
networkx == 3.3